Refactor and fix journey
